+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## (development stage/unreleased)
 
+### Added 
+- Post to Telegram Channel
+
 ## 0.10.2
 ### Changed
 - Delay time 0-15 seconds between retweets